+++ supabase/migrations/20260528112603_update_pages_table_options.sql
@@ -1,0 +1,2 @@
+ALTER TABLE public.pages
+ADD options jsonb;